add canvas for drawing on image

+++ app/src/main/java/com/example/slides_controller/app/Message.java
@@ -5,17 +5,17 @@ import java.util.ArrayList;
 
 public class Message implements Serializable {
     private int operation;
-    private ArrayList<Integer> line_x;
-    private ArrayList<Integer> line_y;
+    private ArrayList<Float> line_x;
+    private ArrayList<Float> line_y;
     private byte[] image;
 
     public Message() {
-        line_x = new ArrayList<Integer>();
-        line_y = new ArrayList<Integer>();
+        line_x = new ArrayList<Float>();
+        line_y = new ArrayList<Float>();
         operation = 0;
     }
 
-    public void addPoint(int x, int y) {
+    public void addPoint(float x, float y) {
         line_x.add(x);
         line_y.add(y);
     }
@@ -28,11 +28,11 @@ public class Message implements Serializable {
         this.operation = operation;
     }
 
-    public ArrayList<Integer> getLine_x() {
+    public ArrayList<Float> getLine_x() {
         return line_x;
     }
 
-    public ArrayList<Integer> getLine_y() {
+    public ArrayList<Float> getLine_y() {
         return line_y;
     }
